requirements.txt
# FastAPI & Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# AI & Search

duckduckgo-search>=6.0.0

# HTTP Client
httpx>=0.27.0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Rate Limiting
slowapi>=0.1.9

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
